[opt](build) Prepare cutting the DataVariants instantiation amplifiers
Pure-additive / behavior-preserving preparation so that the hash-table
variant machinery (exec/common/{agg,join,set,distinct_agg}_utils.h) can be
cut out of exec/pipeline/dependency.h and rec_cte_shared_state.h next.
Those two headers hold non-template in-class inline bodies (SharedState
constructors/destructors/close paths, three std::visit dispatches) that
reference the full variant types, and such bodies are semantically
analyzed when the header is parsed: every one of the ~128 TUs that
transitively include dependency.h instantiates the whole
AggregatedDataVariants/JoinDataVariants/SetDataVariants surface at a flat
~0.85 CPU s per TU (~130 CPU s total), whether it uses them or not. No
include edge is removed here:

- dependency.h: move out of line into dependency.cpp everything whose body
  touches a variant -- AggSharedState ctor/dtor/_close_with_serialized_key/
  _close_without_key, BucketedAggSharedState PerInstanceData ctor, dtor,
  _close and _close_one_agg_data, both HashJoinSharedState ctors, and a
  new user-declared SetSharedState ctor/dtor pair that takes over the
  make_unique<SetDataVariants>() default member initializer. All of it is
  per-query setup/teardown, so outlining costs nothing at runtime
- dependency.h: forward-declare the variant structs, spell the members as
  plain std::unique_ptr/std::vector<size_t> instead of the utils-provided
  aliases, redeclare AggregateDataPtr, and directly include the light
  dependencies that used to ride in through the utils chain (core/arena.h,
  exprs/vexpr_fwd.h, util/stopwatch.hpp, cast_set/exception/
  factory_creator, <list>/<queue>/<set>)
- exec/common/join_op_utils.h (new): JoinOpVariants, is_asof_join*, and
  the AsofIndexGroup/AsofIndexVariant family move here out of
  join_utils.h, which re-exports it; dependency.h holds these by value,
  and this light header depends only on thrift enums, pdqsort and std
  containers -- not on the hash tables
- rec_cte_shared_state.h: same treatment for the fourth amplifier; the
  emplace_block std::visit over DistinctDataVariants plus the brpc-heavy
  build_basic_param/send_data_to_targets bodies move to a new
  rec_cte_shared_state.cpp (CMake GLOB picks it up), with user-declared
  ctor/dtor so unique_ptr<DistinctDataVariants> works forward-declared
- seed direct includes ahead of the cut: agg_utils.h into the three agg
  operator headers whose inline code walks method_variant, join_utils.h
  into join_build_sink_operator.h, distinct_agg_utils.h into
  rec_cte_source_operator.cpp, and the utils/template_helpers includes
  into dependency.cpp

Verified with a full -fsyntax-only sweep: 0 failing TU(s) of 1365.
