+++ be/src/exec/operator/aggregation_sink_operator.h
@@ -19,6 +19,7 @@
 
 #include <stdint.h>
 
+#include "exec/common/agg_utils.h"
 #include "exec/operator/operator.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_profile.h"
