+++ be/src/exec/pipeline/dependency.h
@@ -30,17 +30,25 @@
 #include <atomic>
 #include <condition_variable>
 #include <functional>
+#include <list>
 #include <memory>
 #include <mutex>
+#include <queue>
+#include <set>
 #include <thread>
 #include <utility>
 
+#include "common/cast_set.h"
 #include "common/config.h"
+#include "common/exception.h"
+#include "common/factory_creator.h"
 #include "common/logging.h"
 #include "common/thread_safety_annotations.h"
+#include "core/arena.h"
 #include "core/block/block.h"
 #include "core/types.h"
 #include "exec/common/agg_utils.h"
+#include "exec/common/join_op_utils.h"
 #include "exec/common/join_utils.h"
 #include "exec/common/set_utils.h"
 #include "exec/operator/data_queue.h"
@@ -48,13 +56,23 @@
 #include "exec/sort/partition_sorter.h"
 #include "exec/sort/sorter.h"
 #include "exec/spill/spill_file.h"
+#include "exprs/vexpr_fwd.h"
 #include "runtime/runtime_profile_counter_names.h"
 #include "util/brpc_closure.h"
 #include "util/stack_util.h"
+#include "util/stopwatch.hpp"
 
 namespace doris {
 class AggFnEvaluator;
 class VSlotRef;
+// Heavy hash-table variant machinery (exec/common/*_utils.h) is referenced
+// through pointers only; keep it out of this header's parse cost.
+struct AggregatedDataVariants;
+struct AggregateDataContainer;
+struct BucketedAggDataVariants;
+struct JoinDataVariants;
+struct SetDataVariants;
+using AggregateDataPtr = char*;
 } // namespace doris
 
 namespace doris {
@@ -293,14 +311,11 @@ struct RuntimeFilterTimerQueue {
 struct AggSharedState : public BasicSharedState {
     ENABLE_FACTORY_CREATOR(AggSharedState)
 public:
-    AggSharedState() { agg_data = std::make_unique<AggregatedDataVariants>(); }
-    ~AggSharedState() override {
-        if (!probe_expr_ctxs.empty()) {
-            _close_with_serialized_key();
-        } else {
-            _close_without_key();
-        }
-    }
+    // Defined in dependency.cpp: the bodies touch AggregatedDataVariants'
+    // full variant machinery, which every includer would otherwise
+    // instantiate at parse time.
+    AggSharedState();
+    ~AggSharedState() override;
 
     Status reset_hash_table();
 
@@ -312,7 +327,7 @@ public:
     // 2nd phase: is_merge=false, maybe have multiple exprs.
     static int get_slot_column_id(const AggFnEvaluator* evaluator);
 
-    AggregatedDataVariantsUPtr agg_data = nullptr;
+    std::unique_ptr<AggregatedDataVariants> agg_data;
     std::unique_ptr<AggregateDataContainer> aggregate_data_container;
     std::vector<AggFnEvaluator*> aggregate_evaluators;
     // group by k1,k2
@@ -323,7 +338,7 @@ public:
     size_t total_size_of_aggregate_states = 0;
     size_t align_aggregate_states = 1;
     /// The offset to the n-th aggregate function in a row of aggregate functions.
-    Sizes offsets_of_aggregate_states;
+    std::vector<size_t> offsets_of_aggregate_states;
     std::vector<size_t> make_nullable_keys;
 
     bool agg_data_created_without_key = false;
@@ -398,40 +413,8 @@ public:
 private:
     MutableColumns _get_keys_hash_table();
 
-    void _close_with_serialized_key() {
-        std::visit(Overload {[&](std::monostate& arg) -> void {
-                                 // Do nothing
-                             },
-                             [&](auto& agg_method) -> void {
-                                 if (use_simple_count) {
-                                     // Inline count: mapped slots hold UInt64,
-                                     // not real agg state pointers. Skip destroy.
-                                     return;
-                                 }
-                                 auto& data = *agg_method.hash_table;
-                                 data.for_each_mapped([&](auto& mapped) {
-                                     if (mapped) {
-                                         _destroy_agg_status(mapped);
-                                         mapped = nullptr;
-                                     }
-                                 });
-                                 if (data.has_null_key_data()) {
-                                     _destroy_agg_status(
-                                             data.template get_null_key_data<AggregateDataPtr>());
-                                 }
-                             }},
-                   agg_data->method_variant);
-    }
-
-    void _close_without_key() {
-        //because prepare maybe failed, and couldn't create agg data.
-        //but finally call close to destory agg data, if agg data has bitmapValue
-        //will be core dump, it's not initialized
-        if (agg_data_created_without_key) {
-            _destroy_agg_status(agg_data->without_key);
-            agg_data_created_without_key = false;
-        }
-    }
+    void _close_with_serialized_key();
+    void _close_without_key();
     void _destroy_agg_status(AggregateDataPtr data);
 };
 
@@ -460,22 +443,17 @@ struct BucketedAggSharedState : public BasicSharedState {
     ENABLE_FACTORY_CREATOR(BucketedAggSharedState)
 public:
     BucketedAggSharedState() = default;
-    ~BucketedAggSharedState() override { _close(); }
+    ~BucketedAggSharedState() override; // defined in dependency.cpp (variant machinery)
 
     /// Per-instance data. One per sink pipeline instance.
     /// Each instance has 256 bucket hash tables + 1 shared arena.
     struct PerInstanceData {
         /// 256 per-bucket hash tables. Each bucket has its own BucketedAggDataVariants.
         /// Uses PHHashMap<StringRef> for string keys instead of StringHashMap.
-        std::vector<BucketedAggDataVariantsUPtr> bucket_agg_data;
-        ArenaUPtr arena;
+        std::vector<std::unique_ptr<BucketedAggDataVariants>> bucket_agg_data;
+        std::unique_ptr<Arena> arena;
 
-        PerInstanceData() : arena(std::make_unique<Arena>()) {
-            bucket_agg_data.resize(BUCKETED_AGG_NUM_BUCKETS);
-            for (auto& p : bucket_agg_data) {
-                p = std::make_unique<BucketedAggDataVariants>();
-            }
-        }
+        PerInstanceData(); // defined in dependency.cpp (variant machinery)
     };
 
     /// Per-bucket merge state for pipelined source-side processing.
@@ -514,7 +492,7 @@ public:
     VExprContextSPtrs probe_expr_ctxs;
     size_t total_size_of_aggregate_states = 0;
     size_t align_aggregate_states = 1;
-    Sizes offsets_of_aggregate_states;
+    std::vector<size_t> offsets_of_aggregate_states;
     std::vector<size_t> make_nullable_keys;
 
     std::atomic<size_t> input_num_rows {0};
@@ -563,43 +541,8 @@ private:
     std::once_flag _init_once;
     Status _init_status;
 
-    void _close() {
-        for (auto& inst : per_instance_data) {
-            for (auto& bucket_data : inst.bucket_agg_data) {
-                _close_one_agg_data(*bucket_data);
-            }
-        }
-    }
-
-    void _close_one_agg_data(BucketedAggDataVariants& agg_data) {
-        std::visit(
-                Overload {[&](std::monostate& arg) -> void {
-                              // Do nothing
-                          },
-                          [&](auto& agg_method) -> void {
-                              if (use_simple_count) {
-                                  // simple_count: mapped slots hold UInt64 counters,
-                                  // not real agg state pointers. Skip destroy.
-                                  return;
-                              }
-                              auto& data = *agg_method.hash_table;
-                              data.for_each_mapped([&](auto& mapped) {
-                                  if (mapped) {
-                                      _destroy_agg_status(mapped);
-                                      mapped = nullptr;
-                                  }
-                              });
-                              if constexpr (std::is_assignable_v<decltype(data.has_null_key_data()),
-                                                                 bool>) {
-                                  if (data.has_null_key_data()) {
-                                      _destroy_agg_status(
-                                              data.template get_null_key_data<AggregateDataPtr>());
-                                  }
-                              }
-                          }},
-                agg_data.method_variant);
-    }
-
+    void _close();
+    void _close_one_agg_data(BucketedAggDataVariants& agg_data);
     void _destroy_agg_status(AggregateDataPtr data);
 };
 
@@ -712,16 +655,9 @@ struct JoinSharedState : public BasicSharedState {
 
 struct HashJoinSharedState : public JoinSharedState {
     ENABLE_FACTORY_CREATOR(HashJoinSharedState)
-    HashJoinSharedState() {
-        hash_table_variant_vector.push_back(std::make_shared<JoinDataVariants>());
-    }
-    HashJoinSharedState(int num_instances) {
-        source_deps.resize(num_instances, nullptr);
-        hash_table_variant_vector.resize(num_instances, nullptr);
-        for (int i = 0; i < num_instances; i++) {
-            hash_table_variant_vector[i] = std::make_shared<JoinDataVariants>();
-        }
-    }
+    // Defined in dependency.cpp (they materialize JoinDataVariants).
+    HashJoinSharedState();
+    HashJoinSharedState(int num_instances);
     std::shared_ptr<Arena> arena = std::make_shared<Arena>();
 
     const std::vector<TupleDescriptor*> build_side_child_desc;
@@ -794,6 +730,11 @@ public:
 struct SetSharedState : public BasicSharedState {
     ENABLE_FACTORY_CREATOR(SetSharedState)
 public:
+    // Defined in dependency.cpp: constructing/destroying SetDataVariants
+    // needs the full variant machinery.
+    SetSharedState();
+    ~SetSharedState() override;
+
     /// default init
     Block build_block; // build to source
     //record element size in hashtable
@@ -804,9 +745,8 @@ public:
 
     //// shared static states (shared, decided in prepare/open...)
 
-    /// init in setup_local_state
-    std::unique_ptr<SetDataVariants> hash_table_variants =
-            std::make_unique<SetDataVariants>(); // the real data HERE.
+    /// init in setup_local_state (allocated in the constructor)
+    std::unique_ptr<SetDataVariants> hash_table_variants; // the real data HERE.
     std::vector<bool> build_not_ignore_null;
 
     // The SET operator's child might have different nullable attributes.
